requirements.txt
beautifulsoup4==4.11.1
bleach==5.0.1
docutils==0.19
pipm==22.2.3
pkginfo==1.9.2
Pygments==2.13.0
readme-renderer==37.3
requests-toolbelt==0.10.1
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
webencodings==0.5.1
charset-normalizer==2.1.1
urllib3==1.26.13
idna==3.4
requests==2.28.1
certifi==2022.12.7
load-dotenv==0.1.0
python-dotenv==0.21.0
click==8.1.3
